Style and bug fixes
